Allows for the consumer of the log to choose any kind of formatter
