+++ Serilog.Sinks.Network.Test/Serilog.Sinks.Network.Test.csproj
@@ -73,7 +73,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="TcpTests.cs" />
+    <Compile Include="TCPTests.cs" />
     <Compile Include="UDPTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
