+++ Serilog.Sinks.Network/NetworkLoggerConfigurationExtensions.cs
@@ -4,6 +4,8 @@ using System.Net;
 using Serilog.Configuration;
 using Serilog.Debugging;
 using Serilog.Events;
+using Serilog.Formatting;
+using Serilog.Sinks.Network.Formatters;
 using Serilog.Sinks.Network.Sinks.TCP;
 using Serilog.Sinks.Network.Sinks.UDP;
 
@@ -18,20 +20,20 @@ namespace Serilog.Sinks.Network
             this LoggerSinkConfiguration loggerConfiguration,
             string uri,
             int port,
+            ITextFormatter textFormatter = null,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
         {
-            var sink = new UDPSink(ResolveAddress(uri), port);
-            return loggerConfiguration.Sink(sink, restrictedToMinimumLevel);
+            return UDPSink(loggerConfiguration, ResolveAddress(uri), port, textFormatter, restrictedToMinimumLevel);
         }
 
         public static LoggerConfiguration UDPSink(
             this LoggerSinkConfiguration loggerConfiguration,
             IPAddress ipAddress,
             int port,
+            ITextFormatter textFormatter = null,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
         {
-            var sink = new UDPSink(ipAddress, port);
-
+            var sink = new UDPSink(ipAddress, port, textFormatter ?? new LogstashJsonFormatter());
             return loggerConfiguration.Sink(sink, restrictedToMinimumLevel);
         }
 
@@ -39,10 +41,10 @@ namespace Serilog.Sinks.Network
             this LoggerSinkConfiguration loggerConfiguration,
             IPAddress ipAddress,
             int port,
+            ITextFormatter textFormatter = null,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
         {
-            var sink = new TCPSink(ipAddress, port);
-
+            var sink = new TCPSink(ipAddress, port, textFormatter ?? new LogstashJsonFormatter());
             return loggerConfiguration.Sink(sink, restrictedToMinimumLevel);
         }
 
@@ -50,18 +52,19 @@ namespace Serilog.Sinks.Network
             this LoggerSinkConfiguration loggerConfiguration,
             string host,
             int port,
+            ITextFormatter textFormatter = null,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
         {
-            var sink = new TCPSink(BuildUri(string.Format("{0}:{1}", host, port)));
-            return loggerConfiguration.Sink(sink, restrictedToMinimumLevel);
+            return TCPSink(loggerConfiguration, $"{host}:{port}", textFormatter, restrictedToMinimumLevel);
         }
 
         public static LoggerConfiguration TCPSink(
             this LoggerSinkConfiguration loggerConfiguration,
             string uri,
+            ITextFormatter textFormatter = null,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
         {
-            var sink = new TCPSink(BuildUri(uri));
+            var sink = new TCPSink(BuildUri(uri), textFormatter ?? new LogstashJsonFormatter());
             return loggerConfiguration.Sink(sink, restrictedToMinimumLevel);
         }
 
@@ -97,7 +100,7 @@ namespace Serilog.Sinks.Network
             }
         }
 
-        static Uri BuildUri(string s)
+        private static Uri BuildUri(string s)
         {
             Uri uri;
             try {
